174: move meta instantiation into meta provider
